+++ HISTORY.rst
@@ -12,6 +12,7 @@ released ...
 - breaking: python 2.6 is no longer supported
 - breaking: webclient.get_registered_devices has a slightly different schema
 - fix Webclient authentication and get_stream_urls
+- fix MusicManager uploading: Google shut down the rupio endpoint
 - fix certificate validation
 - fix album artist metadata not being upload