+++ HISTORY.rst
@@ -7,7 +7,7 @@ As of 1.0.0, `semantic versioning <http://semver.org/>`__ is used.
 
 7.0.0
 +++++
-released ...
+released 2015-09-19
 
 - breaking: python 2.6 is no longer supported
 - breaking: webclient.get_registered_devices has a slightly different schema